Add word check and styling, word list validation, and display updates
A lot happening here, but now, basically the game will apply a class
to guessed letters to style them based on their inclusion in the word
and if their position is correct. A check was also added upon hitting enter
to make sure that the guessed word is actually in the word list. This also
showed me that the word list probably needs to be expanded, too.